Moves up IntervalManager code to the Core and updates Eclipse code
Refs #219
